+++ packages/featureserver/src/geometry/extents.js
@@ -1,7 +1,12 @@
-module.exports = function (input) {
+module.exports = function (input, spatialReference) {
   let coords
   if (!input) return undefined
-  if (!isNaN(input.xmin)) return input
+  if (!isNaN(input.xmin)) {
+    return {
+      ...input,
+      spatialReference
+    }
+  }
   if (Array.isArray(input)) {
     if (Array.isArray(input[0])) coords = input
     else coords = [[input[0], input[1]], [input[2], input[3]]]
@@ -13,9 +18,6 @@ module.exports = function (input) {
     ymin: coords[0][1],
     xmax: coords[1][0],
     ymax: coords[1][1],
-    spatialReference: {
-      wkid: 4326,
-      latestWkid: 4326
-    }
+    spatialReference
   }
 }