+++ packages/featureserver/src/route.js
@@ -79,9 +79,9 @@ function execInfo (req, res, method, geojson) {
     if (/\/rest\/info$/i.test(url)) {
       info = FsInfo.restInfo(geojson)
     } else if (/\/FeatureServer$/i.test(url) || /\/FeatureServer\/info$/i.test(url) || /\/FeatureServer\/($|\?)/.test(url)) {
-      info = FsInfo.serverInfo(geojson, req.params)
+      info = FsInfo.serverInfo(geojson, req)
     } else if (/\/FeatureServer\/\d+$/i.test(url) || /\/FeatureServer\/\d+\/info$/i.test(url) || /\/FeatureServer\/\d+\/($|\?)/.test(url)) {
-      info = FsInfo.layerInfo(geojson, req.params)
+      info = FsInfo.layerInfo(geojson, req)
     } else if (/\/FeatureServer\/layers$/i.test(url)) {
       info = FsInfo.layersInfo(geojson, req.query)
     } else if (/\/FeatureServer/i.test(url)) {